+++ packages/stream_chat_flutter/lib/src/message_widget.dart
@@ -92,6 +92,8 @@ class MessageWidget extends StatefulWidget {
     this.userAvatarBuilder,
     this.editMessageInputBuilder,
     this.textBuilder,
+    this.bottomRowBuilder,
+    this.deletedBottomRowBuilder,
     this.onReturnAction,
     Map<String, AttachmentBuilder>? customAttachmentBuilders,
     this.readList,
@@ -275,6 +277,12 @@ class MessageWidget extends StatefulWidget {
   /// Function called on long press
   final void Function(BuildContext, Message)? onMessageActions;
 
+  /// Widget builder for building a bottom row below the message
+  final Widget Function(BuildContext, Message)? bottomRowBuilder;
+
+  /// Widget builder for building a bottom row below a deleted message
+  final Widget Function(BuildContext, Message)? deletedBottomRowBuilder;
+
   /// Widget builder for building user avatar
   final Widget Function(BuildContext, User)? userAvatarBuilder;
 
@@ -410,6 +418,8 @@ class MessageWidget extends StatefulWidget {
     Widget Function(BuildContext, Message)? editMessageInputBuilder,
     Widget Function(BuildContext, Message)? textBuilder,
     Widget Function(BuildContext, Message)? usernameBuilder,
+    Widget Function(BuildContext, Message)? bottomRowBuilder,
+    Widget Function(BuildContext, Message)? deletedBottomRowBuilder,
     void Function(BuildContext, Message)? onMessageActions,
     Message? message,
     MessageTheme? messageTheme,
@@ -463,6 +473,9 @@ class MessageWidget extends StatefulWidget {
             editMessageInputBuilder ?? this.editMessageInputBuilder,
         textBuilder: textBuilder ?? this.textBuilder,
         usernameBuilder: usernameBuilder ?? this.usernameBuilder,
+        bottomRowBuilder: bottomRowBuilder ?? this.bottomRowBuilder,
+        deletedBottomRowBuilder:
+            deletedBottomRowBuilder ?? this.deletedBottomRowBuilder,
         onMessageActions: onMessageActions ?? this.onMessageActions,
         message: message ?? this.message,
         messageTheme: messageTheme ?? this.messageTheme,
@@ -782,7 +795,11 @@ class _MessageWidgetState extends State<MessageWidget>
                             bottom:
                                 isPinned && widget.showPinHighlight ? 6.0 : 0.0,
                           ),
-                          child: _bottomRow,
+                          child: widget.bottomRowBuilder?.call(
+                                context,
+                                widget.message,
+                              ) ??
+                              _bottomRow,
                         ),
                       if (isFailedState)
                         Positioned(
@@ -830,22 +847,11 @@ class _MessageWidgetState extends State<MessageWidget>
 
   Widget get _bottomRow {
     if (isDeleted) {
-      final chatThemeData = _streamChatTheme;
-      return Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          StreamSvgIcon.eye(
-            color: chatThemeData.colorTheme.textLowEmphasis,
-            size: 16,
-          ),
-          const SizedBox(width: 8),
-          Text(
-            'Only visible to you',
-            style: chatThemeData.textTheme.footnote
-                .copyWith(color: chatThemeData.colorTheme.textLowEmphasis),
-          ),
-        ],
-      );
+      return widget.deletedBottomRowBuilder?.call(
+            context,
+            widget.message,
+          ) ??
+          const Offstage();
     }
 
     final children = <Widget>[];