+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -42,3 +42,4 @@ export 'src/user_item.dart';
 export 'src/user_list_view.dart';
 export 'src/user_list_view.dart';
 export 'src/utils.dart';
+export 'src/visible_footnote.dart';