+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `MessageListView.paginationLimit`
 - Allow the various ListView widgets to be themed via ThemeData classes
+- Added `bottomRowBuilder` and `deletedBottomRowBuilder` that build a widget below a `MessageWidget`
 
 🔄 Changed